fix: refactoring client adn backend servers code
